pyproject.toml
[project]
name = "ticket-ai-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "SQLAlchemy>=2.0.0",
    "langgraph>=0.6.0",
    "langgraph-checkpoint-postgres>=0.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.7.0",
    "psycopg[binary,pool]>=3.2.1",
    "python-dotenv>=1.0.1",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
